py_asr_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
faster-whisper==1.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
huggingface_hub[hf_xet]